+++ actix/CHANGES.md
@@ -3,8 +3,19 @@
 ## Unreleased - 2021-xx-xx
 ### Added
 * Added `actix::prelude::AsyncResponse` to use async/await syntax in actor handler [#449]
+* Add `fut::try_future::ActorTryFuture`. [#419]
+* Add `fut::try_future::ActorTryFutureExt` trait with `map_ok`, `map_err` and `and_then` combinator. [#419] 
+* Added `actix::prelude::AsyncResponse` to use async/await syntax in actor handler [#449]
 
+[#419]: https://github.com/actix/actix/pull/419
 [#449]: https://github.com/actix/actix/pull/449
+
+
+## 0.11.1 - 2021-03-23
+### Fixed
+* Panics caused by instant cancellation of a spawned future. [#484]
+
+[#484]: https://github.com/actix/actix/pull/484
 
 
 ## 0.11.0 - 2021-03-21